Shop now uses Conjured item
Before this, when you created a new shop with a conjured
item as part of the constructor paramaters, it would
use that object as the item and then had a method for updating
the property values (_conjuredUpdate). After this commit it
creates a Conjured object which has an update method on it
which updates its own state.